requirements.txt
django==3.2.2
gunicorn==20.1.0
djangorestframework==3.12.4
psycopg2-binary==2.9.2
django-storages[google]==1.12.3
django-environ==0.8.1
google-cloud-secret-manager==2.8.0
google-auth==2.3.3